Fixing the way the SparkSession structure is provided to tests.
We were using the loan pattern, but tests are failing on Travis because the Spark
Session is being shutdown. [This answer](https://stackoverflow.com/questions/54187133/unit-tests-using-spark-session-sparkcontext-was-shut-down)
suggests the problem is linked to starting and stopping multiple spark contexts on the
same JVM.

Refactor doesn't stop the SparkSession at the end, but it should be cleaned up automatically
when the test suite ends.