Add ETLSessionContext to EtlSparkUnitTest
Some methods require that we pass in an instances of ETLSessionContext. This is created with a custom Spark session for testing and uses the default configuration. If the configTest passes, there should be no problem loading an ETLSessionContext.

It's possible to override the configuration by adding a custom application.conf file in the test/resources directory or overriding the system properties.