Cancer biomarkers rewrite (#147)
* Rewrite cancer biomarker.

CancerBiomarker was previously generated by the data pipeline and then
modified slightly by the ETL. This rewrite makes the data pipeline
redundant.

Two static files have been extracted from the DP and are now in the
resources directory, although they should be made available through PIS.

This step relies on the target step to map the gene name to an ENSG ID.

* Add ETLSessionContext to EtlSparkUnitTest

Some methods require that we pass in an instances of ETLSessionContext. This is created with a custom Spark session for testing and uses the default configuration. If the configTest passes, there should be no problem loading an ETLSessionContext.

It's possible to override the configuration by adding a custom application.conf file in the test/resources directory or overriding the system properties.

* Add Test for CancerBiomarker

* Update static test input files to jsonl

* Update Main to use new CancerBiomarkers step.

* Delete old CancerBiomarker implementation.

* Update Cancer Biomarker input configuration

- Add logging to Cancer Biomarker.