Target step updates: alternative gene and gene ontology (#152)
* Update `alternativeGenes` on Target.

Fixed logic bug where alternative genes on non-canonical chromosomes which shared an approved symbol with a canonical gene were not properly grouped.

Relates to opentargets/platform#1537

* Update GeneOntology on Target to improve mapping

GeneOntology needs to be mapped to a uniport protein id. These protein ids are extracted from the Uniprot intput, and can be labelled 'uniprot_trembl' or 'uniport_swissprot'. This change ensures we collect all the values we need for the merge.