1678 add pathways to target (#168)
* Update documentation: link between Reactome and Target

* Add pathways to Target

* Removing unused method from Target

This is needed generally to map to datasets which don't include
Ensembl Ids, but can be included in a more appropriate PR.

* Prevent running of tests twice in Travis.

We run the tests on line 8. The `assembly` command calls them all
again which is unnecessary as the build will have failed before then
if there is a problem.