Add target validation step to ETL. (#215)
* Add target validation step to ETL.

Target validation step is used for cases when we only want to check that
all the rows in a dataset are also present in the target dataset. In
this way we ensure that ETL outputs link up nicely and we don't have
polluted datasets.

Resolves opentargets/platform#1798

* Update target validation to use semi-join instead of inner.

* Clean target validation of unnecessary 'optimisation'

We don't need to preselect the id column on target. According to the
generated physical plan (`cleanedDf.explain()`) Spark is only loading
the `id` column of `targetDf`. Since we're using a 'left_semi' join now
we don't need to worry about column name clashes either.