Add canonical transcripts to Target object.
There are 57 targets which do not have canonical transcripts. These are
not present in the gencode input file. As a workaround, in cases where
there is only a single transcript that should also be listed as the
canonical one. In cases where there are multiple transcripts leave the
CanonicalTranscript as null.