Update Metadata generation.
Updates metadata generation so that all outputs are provided with
metadata.

Metadata is now saved in the directory
`common.output-base-path`/metadata . Metadata for each dataset is
appended to this directory, so if we run the pipeline multiple times in
the same place without cleaning up after ourselves we potentially will
have multiple metadata files for the same input.

Resolves opentargets/platform#2780