Split GCP dependencies for modules and shade shapeless.
Shapeless is used by the pureconfig library and Spark. Because we list
Spark as a provided dependency we 'inherit' the one on the cluster. The
solution comes from 'https://stackoverflow.com/questions/42568010/nosuchmethoderror-in-shapeless-seen-only-in-spark'

Split dependencies between modules ETL and Workflow. There is some
duplication, but it's clearer where the dependencies are used.