Remove old scripts.
These are a bunch of work-in-progress/proof-of-concept scripts which are
not used anywhere. Most of them were working towards the literature
pipeline.

The main exception is the target safety script which was used to
deprecate some old spreadsheets. Since we've been using the new data now
for several releases, it does not retain any value.