fix(MechanismOfAction): do not exclude moas with an empty list of targets
`targets` is a list of all EnsemblIDs that drug is targeting to. It is not correct to drop the records where this list is empty, because in many cases these refer to drugs that target a non human protein. Therefore, we just want to check they either an ensembl ID or a target name is present. More context here https://github.com/opentargets/issues/issues/3005#issuecomment-1697290163